refactor(common): promote createDebouncedFn to @sero-ai/common
Move the plain (non-hook) debounce factory into the shared package so
plugins and stores reuse one implementation with a real cancel(), instead
of hand-rolled setTimeout copies. The desktop useDebouncedCallback hook
re-exports it for existing import sites.
